requirements.txt
selenium==4.13.0
python-dotenv==1.0.0